disabled registration

+++ routes/web.php
@@ -20,16 +20,20 @@ Route::get('/', function () {
   return view('welcome');
 });
 
-Route::get('/wines', [WineController::class, 'index']);
+// middleware is going to check if the user is currently logged in, if not redirect.
+Route::get('/wines', [WineController::class, 'index'])->middleware('auth');
 
 // create needs to be above the id route, bc they rank in order of preference
 // it will look at this route literaly and take the create route over the id route
 
 Route::get('/wines/create', [WineController::class, 'create']);
 Route::post('/wines', [WineController::class, 'store']);
-Route::get('/wines/{id}', [WineController::class, 'show']);
-Route::delete('/wines/{id}', [WineController::class, 'destroy']);
+Route::get('/wines/{id}', [WineController::class, 'show'])->middleware('auth');
+Route::delete('/wines/{id}', [WineController::class, 'destroy'])->middleware('auth');
 
 
-Auth::routes();
+Auth::routes([
+  'register' => false
+]);
+
 Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
